requirements.txt
fastapi~=0.115.12
uvicorn~=0.34.2
aiosqlite~=0.21.0
pydantic~=2.11.4
python-jose~=3.4.0
SQLAlchemy~=2.0.41
passlib~=1.7.4